+++ src/components/Auth/AuthInput/AuthInputBox/AuthInputItem/index.ts
@@ -1,0 +1,6 @@
+import AuthInputWrapper from './AuthInputWrapper';
+import AuthInput from './AuthInput';
+import ValidifyCheck from './ValidifyCheck';
+import ErrorMessage from './ErrorMessage';
+
+export { AuthInput, AuthInputWrapper, ErrorMessage, ValidifyCheck };
