+++ src/components/Auth/index.ts
@@ -1,10 +1,12 @@
-import AuthInput from './AuthInput/AuthInput';
+import {
+  AuthInput,
+  ErrorMessage,
+} from './AuthInput/AuthInputBox/AuthInputItem';
 import {
   AuthEmailInputBox,
   AuthPwInputBox,
   AuthPwCheckInputBox,
 } from './AuthInput/AuthInputBox';
-import ErrorMessage from './ErrorMessage';
 import KakaoLoginButton from './KakaoLoginButton';
 import SignupButton from './SignupButton';
